2/19/2023 Afternoon Changes
---Created an inventory system, that updates as the player picks up items.
It allows for stackable and unstackable items, and prevents the players from going over the maximum number of inventory slots.
The ability to drop items is working on the UI side, but I still need to make the player ACTUALLY drop the items in-game.

---KNOWN ISSUE: Selecting the inventory buttons to drop items is barely working on a gamepad, with no clear reason as to why.
I may have to abandon my inventory system entirely, in favor of a plugin or Marketplace asset that provides better results AND better gamepad support.